requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.6,<3.0
requests>=2.31,<3.0
PyYAML>=6.0,<7.0
Pillow>=10.0,<12.0
transformers>=4.40,<5.0
torch>=2.2
sentencepiece>=0.2.0,<1.0
protobuf>=4.25,<6.0
